scripts/requirements.txt
oracledb>=2.0
google-cloud-bigquery>=3.20
pandas>=2.2
pyarrow>=15.0
